requirements.txt
python-dotenv>=1.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0
pandas>=2.0.0
matplotlib>=3.8.0
Pillow>=10.0.0
modal>=0.73
